Trim emailAddress input whitespace
For some reason, when setting the input element value for
the email address and email address validation inputs, two
whitespace characters are introduced.

These whitespace characters (leading) are trimmed for
the email address input but not the validation input.

Therefore, this fix solves an error related to the email
address inputs not matching after submission.